move File to java dir
